+++ test/xema/list_test.exs
@@ -102,9 +102,7 @@ defmodule Xema.ListTest do
       assert {:error,
               %ValidationError{
                 reason: %{
-                  items: [
-                    {2, %{type: :integer, value: "foo"}}
-                  ]
+                  items: %{2 => %{type: :integer, value: "foo"}}
                 }
               } = error} = validate(schema, [1, 2, "foo"])
 
@@ -124,10 +122,10 @@ defmodule Xema.ListTest do
                :error,
                %ValidationError{
                  reason: %{
-                   items: [
-                     {0, %{type: :string, value: 1}},
-                     {1, %{type: :string, value: 2}}
-                   ]
+                   items: %{
+                     0 => %{type: :string, value: 1},
+                     1 => %{type: :string, value: 2}
+                   }
                  }
                } = error
              } = validate(schema, [1, 2, "foo"])
@@ -231,7 +229,7 @@ defmodule Xema.ListTest do
                :error,
                %ValidationError{
                  reason: %{
-                   items: [{1, %{type: :number, value: "bar"}}]
+                   items: %{1 => %{type: :number, value: "bar"}}
                  }
                } = error
              } = validate(schema, ["foo", "bar"])
@@ -242,7 +240,7 @@ defmodule Xema.ListTest do
                :error,
                %ValidationError{
                  reason: %{
-                   items: [{0, %{value: "x", min_length: 3}}]
+                   items: %{0 => %{value: "x", min_length: 3}}
                  }
                } = error
              } = validate(schema, ["x", 33])
@@ -254,7 +252,7 @@ defmodule Xema.ListTest do
       assert {
                :error,
                %ValidationError{
-                 reason: %{items: [{0, %{value: "x", min_length: 3}}]}
+                 reason: %{items: %{0 => %{value: "x", min_length: 3}}}
                } = error
              } = validate(schema, ["x", 33, 7])
 
@@ -289,7 +287,7 @@ defmodule Xema.ListTest do
       assert {
                :error,
                %ValidationError{
-                 reason: %{items: [{2, %{additional_items: false}}]}
+                 reason: %{items: %{2 => %{additional_items: false}}}
                } = error
              } = validate(schema, ["foo", 42, "add"])
 
@@ -320,7 +318,7 @@ defmodule Xema.ListTest do
                :error,
                %ValidationError{
                  reason: %{
-                   items: [{2, %{type: :string, value: 13}}]
+                   items: %{2 => %{type: :string, value: 13}}
                  }
                } = error
              } = validate(schema, [11, "twelve", 13])